requirement.txt
fastapi
uvicorn
pydantic-settings

sqlalchemy
asyncpg
alembic

python-multipart

pytesseract 
pdf2image 
python-docx 
opencv-python-headless 
spacy
pymupdf